Added public LocationNode method to stack the node
